requirements.txt
numpy
Pillow
scikit-learn
torch
torchvision
tqdm
transformers
flair
